Install lodash Lodash is a 3rd party library that provides many methods. The methods can be accessed by prefixing the underscore "_". It is installed by the command: "npm i --save lodash" Example Usage const _ = require('lodash');
const num = _.random(1, 20);

console.log(num); //will log a random number to the
console

Note that nodemon (a live reload server) was
installed globally so I did not need to install it
again locally.

Also note that when a 3rd party package is installed
for the first time, a 'node_modules' folder appears
automatically. This folder contains all the
3rd party resources for the installed packages